Modernization...

+++ src/ifmetadata.h
@@ -191,7 +191,7 @@ enum IFMDErrorType {
 struct IFMDError {
 	enum IFMDSeverity severity;
 	enum IFMDErrorType type;
-	int lineNumber;
+	long lineNumber;
 
 	char* moreText;
 };
